fix(ci): cast SFTP/FTP port to int in filesystems config
The release-readiness merge added ext-ftp to the Laravel image, so the
CloudStorageConfigTest sftp/ftp disk-resolution cases now actually run
instead of being skipped. env('SFTP_PORT')/env('FTP_PORT') return strings
in the container, and Flysystem's SftpConnectionProvider/FtpConnectionOptions
require an int $port, so both cases threw TypeError and reddened CI.

Cast at the config boundary — coerces whether the env yields a string or an
int, in CI and in any real deployment. All 11 CloudStorageConfigTest pass.

+++ archive-laravel/config/filesystems.php
@@ -97,7 +97,7 @@ return [
         'sftp' => [
             'driver' => 'sftp',
             'host' => env('SFTP_HOST'),
-            'port' => env('SFTP_PORT', 22),
+            'port' => (int) env('SFTP_PORT', 22),
             'username' => env('SFTP_USERNAME'),
             'password' => env('SFTP_PASSWORD'),
             'privateKey' => env('SFTP_PRIVATE_KEY'),
@@ -110,7 +110,7 @@ return [
         'ftp' => [
             'driver' => 'ftp',
             'host' => env('FTP_HOST'),
-            'port' => env('FTP_PORT', 21),
+            'port' => (int) env('FTP_PORT', 21),
             'username' => env('FTP_USERNAME'),
             'password' => env('FTP_PASSWORD'),
             'root' => env('FTP_ROOT', ''),
